Wallhug turn right 1st if there exist unexplored during left turn(wallhug type1) (old wallhug is type 4)
enable auto calibrate option
increase greedy explore>distance (wallhug type1)

+++ AlgorithmSimulator/src/application/MainController.java
@@ -424,7 +424,7 @@ public class MainController extends FXMLController  implements Initializable, Ro
 	}
 	@FXML
 	public void onTerminateRateTextField() {
-		Exploration.setAutoTerminate_explore_rate(Float.parseFloat(terminateRateTextField.getText()));
+		Exploration.setAutoTerminate_explore_rate(Integer.parseInt(terminateRateTextField.getText()));
 	}
 	RPiInterface rpi = null;
 	@FXML
